requirements.txt
torch>=1.10.0
numpy>=1.19.0
pandas>=1.3.0
matplotlib>=3.4.0
tensorflow>=2.8.0
# tensorflow-federated==0.20.0
# jaxlib>=0.4.13
gymnasium>=0.26.0
networkx>=2.6.0
tqdm>=4.62.0
scikit-learn>=1.0.0
pytest>=6.2.5
wandb>=0.12.0
seaborn>=0.11.0
py-ecc>=6.0.0
cryptography>=37.0.0
